refactor: refactor models and add exceptions to zipcode and invalid coupon
