--- app/infra/base.py
@@ -1,10 +1,0 @@
-from app.infra.models.role import Role
-from app.infra.models.transaction import (
-    CreditCardFeeConfig,
-    Payment,
-    Transaction,
-)
-from app.infra.models.uploadedimage import UploadedImage
-from app.infra.models.users import Address, User
-
-from app.infra.models.base import Base
