+++ app/infra/endpoints/direct_sales.py
@@ -1,7 +1,6 @@
 from fastapi import APIRouter, Depends
 from sqlalchemy.orm import Session
 
-from domains import domain_order, domain_payment
 from app.infra import deps
 from schemas.order_schema import CheckoutReceive, ProductSchema
 
@@ -15,19 +14,13 @@ async def get_product(
     uri: None,
 ) -> None:
     """Get product."""
-    try:
-        return domain_order.get_product(db, uri)
-    except Exception:
-        raise
+    ...
 
 
 @direct_sales.get('/upsell/{id}', status_code=200)
 async def get_upsell_products() -> None:
     """Get upsell products."""
-    try:
-        return True
-    except Exception:
-        raise
+    ...
 
 
 @direct_sales.post('/checkout', status_code=201)
@@ -37,21 +30,7 @@ def checkout(
     data: CheckoutReceive,
 ) -> None:
     """Checkout."""
-    try:
-        checkout_data = data.dict().get('transaction')
-        affiliate = data.dict().get('affiliate')
-        cupom = data.dict().get('cupom')
-        from loguru import logger
-
-        logger.info(checkout_data)
-        return domain_payment.process_checkout(
-            db=db,
-            checkout_data=checkout_data,
-            affiliate=affiliate,
-            cupom=cupom,
-        )
-    except Exception:
-        raise
+    ...
 
 
 @direct_sales.post('/create-product', status_code=201)
@@ -61,5 +40,4 @@ async def create_product(
     product_data: ProductSchema,
 ) -> None:
     """Create product."""
-    product = domain_order.create_product(db=db, product_data=product_data)
-    return ProductSchema.from_orm(product)
+    ...
