+++ app/infra/endpoints/product.py
@@ -4,7 +4,7 @@ from loguru import logger
 from sqlalchemy.orm import Session
 from app.entities.product import ProductInDB
 
-from domains import domain_order
+from app.order import services
 from app.infra import deps
 from app.infra.deps import get_db
 from payment.schema import ProductSchema
@@ -30,7 +30,7 @@ def create_product(
     product_data: ProductSchema,
 ) -> ProductSchema:
     """Create product."""
-    return domain_order.create_product(db=db, product_data=product_data)
+    return services.create_product(db=db, product_data=product_data)
 
 
 @product.post('/upload-image/{product_id}', status_code=200)
@@ -41,7 +41,7 @@ def upload_image(
 ) -> None:
     """Upload image."""
     try:
-        return domain_order.upload_image(db, product_id, image)
+        return services.upload_image(db, product_id, image)
     except Exception:
         raise
 
@@ -54,7 +54,7 @@ def upload_image_gallery(
 ) -> None:
     """Upload image gallery."""
     try:
-        return domain_order.upload_image_gallery(product_id, db, imagegallery)
+        return services.upload_image_gallery(product_id, db, imagegallery)
     except Exception:
         raise
 
@@ -63,7 +63,7 @@ def upload_image_gallery(
 def get_images_gallery(uri: str, db: Session = Depends(get_db)) -> None:
     """Get images gallery."""
     try:
-        return domain_order.get_images_gallery(db, uri)
+        return services.get_images_gallery(db, uri)
     except Exception:
         raise
 
@@ -72,7 +72,7 @@ def get_images_gallery(uri: str, db: Session = Depends(get_db)) -> None:
 def delete_image(id: int, db: Session = Depends(get_db)) -> None:
     """Delete image."""
     try:
-        return domain_order.delete_image_gallery(id, db)
+        return services.delete_image_gallery(id, db)
     except Exception:
         raise
 
@@ -81,7 +81,7 @@ def delete_image(id: int, db: Session = Depends(get_db)) -> None:
 def get_product_uri(uri: str, db: Session = Depends(get_db)) -> ProductInDB:
     """GET product uri."""
     try:
-        product = domain_order.get_product(db, uri)
+        product = services.get_product(db, uri)
         if not product:
             raise HTTPException(
                 status_code=status.HTTP_404_NOT_FOUND,
@@ -102,7 +102,7 @@ def put_product(
 ) -> None:
     """Put product."""
     try:
-        return domain_order.put_product(db, id, value)
+        return services.put_product(db, id, value)
     except Exception:
         raise
 
@@ -111,7 +111,7 @@ def put_product(
 def delete_product(id: int, db: Session = Depends(get_db)) -> None:
     """Delete product."""
     try:
-        return domain_order.delete_product(db, id)
+        return services.delete_product(db, id)
     except Exception:
         raise
 
@@ -124,7 +124,7 @@ def get_installments(
 ) -> Any:
     """Get installments."""
     try:
-        return domain_order.get_installments(product_id, db=db)
+        return services.get_installments(product_id, db=db)
     except Exception as e:
         logger.error(f'Erro ao coletar o parcelamento - {e}')
         raise
