+++ app/mail/tasks.py
@@ -30,7 +30,7 @@ def task_mail_order_cancelled(
     db: Session = Depends(get_db),
 ) -> None:
     """Send cancelled email."""
-    logger.info("Start task to send mail order cancelled.")
+    logger.info('Start task to send mail order cancelled.')
     mail_data = MailOrderCancelled(
         mail_to=mail_to,
         order_id=order_id,
@@ -46,7 +46,7 @@ def task_mail_order_processed(
     db: Session = Depends(get_db),
 ) -> None:
     """Send cancelled email."""
-    logger.info("Start task to send mail order processed.")
+    logger.info('Start task to send mail order processed.')
     mail_data = MailOrderProcessed(
         mail_to=mail_to,
         order_id=order_id,
@@ -61,7 +61,7 @@ def task_mail_order_paid(
     db: Session = Depends(get_db),
 ) -> None:
     """Send cancelled email."""
-    logger.info("Start task to send mail order paid.")
+    logger.info('Start task to send mail order paid.')
     logger.info(f'{mail_to}')
     logger.info(f'{order_id}')
     mail_data = MailOrderPaied(
