+++ app/order/uow.py
@@ -4,7 +4,7 @@ from typing import Any
 from sqlalchemy.orm import SessionTransaction
 from app.entities.cart import CartPayment
 from app.infra.custom_decorators import database_uow
-from app.infra.models import order
+from app.infra.models import OrderDB
 from app.user import repository as user_repository
 from app.order import repository as order_repository
 from app.order.entities import OrderDBUpdate
@@ -61,7 +61,7 @@ async def uow_update_paid_order(
     order_update: OrderDBUpdate,
     transaction: SessionTransaction | None,
     bootstrap: Any,
-) -> order.Order:
+) -> OrderDB:
     if not transaction:
         msg = 'Transaction must be provided'
         raise ValueError(msg)
