+++ app/user/repository.py
@@ -1,16 +1,16 @@
 from sqlalchemy import select, update
 
 from sqlalchemy.orm import SessionTransaction
-from app.infra.models import users
+from app.infra import models
 
 
 async def get_user_by_id(
     user_id: int,
     *,
     transaction: SessionTransaction,
-) -> users.User:
+) -> models.UserDB:
     """Get an user by its id."""
-    user_query = select(users.User).where(users.User.user_id == user_id)
+    user_query = select(models.UserDB).where(models.UserDB.user_id == user_id)
     return await transaction.session.scalar(user_query)
 
 
@@ -18,22 +18,24 @@ async def get_user_by_username(
     username: str,
     *,
     transaction: SessionTransaction,
-) -> users.User | None:
+) -> models.UserDB | None:
     """Get an user by its username."""
-    user_query = select(users.User).where(users.User.username == username)
+    user_query = select(models.UserDB).where(
+        models.UserDB.username == username,
+    )
     return await transaction.session.scalar(user_query)
 
 
 async def update_user(
     user_id: int,
     *,
-    user: users.User,
+    user: models.UserDB,
     transaction: SessionTransaction,
-) -> users.User:
+) -> models.UserDB:
     """Update an user."""
     user_query = (
-        update(users.User)
-        .where(users.User.user_id == user_id)
+        update(models.UserDB)
+        .where(models.UserDB.user_id == user_id)
         .values(customer_id=user.customer_id)
     )
     return await transaction.session.execute(user_query)
